refactor: move graphemes helper into text/internal/
It's shared logic used by reverse.ts and truncate.ts but not publicly
exported, matching the rest of text/internal/'s contents.
